[Controls] Changed ProgressBar to use Double instead of Float

+++ nifty-controls-illarion/src/org/illarion/nifty/controls/Progress.java
@@ -29,5 +29,5 @@ public interface Progress extends NiftyControl {
      *
      * @param value the progress value
      */
-    void setProgress(float value);
+    void setProgress(double value);
 }
